server/requirements-full.txt
opencv-python==4.8.1.78
numpy==1.24.3
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0 
faster-whisper
pyannote.audio
transformers
librosa
soundfile
mediapipe>=0.10.0
scipy>=1.10.0

# REVOLUTIONARY AI ENHANCEMENTS
# Multi-Modal Intelligence
face-recognition>=1.3.0
dlib>=19.24.0
mtcnn>=0.1.1
keras-facenet>=0.3.2
emotion-recognition>=1.0.0

# Advanced Audio Analysis
pyaudio>=0.2.11
webrtcvad>=2.0.10
speechbrain>=0.5.0
praat-parselmouth>=0.4.3
python-speech-features>=0.6

# Computer Vision & Visual Intelligence
ultralytics>=8.0.0
detectron2>=0.6
yolov8>=0.0.1
insightface>=0.7.3
facenet-pytorch>=2.5.3

# Natural Language Understanding
spacy>=3.7.0
nltk>=3.8.1
textblob>=0.17.1
vaderSentiment>=3.3.2
sentence-transformers>=2.2.2
bert-score>=0.3.13

# Machine Learning & Deep Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0
tensorflow>=2.13.0
keras>=2.13.1
pytorch-lightning>=2.0.0

# Data Analysis & Visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0

# Real-Time Data & APIs
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
tweepy>=4.14.0
google-api-python-client>=2.90.0
instagrapi>=1.19.0

# Performance & Optimization
numba>=0.57.0
cython>=0.29.0
psutil>=5.9.0
memory-profiler>=0.60.0
line-profiler>=4.0.0

# Database & Caching
redis>=4.6.0
sqlite3
pymongo>=4.4.0

# Async & Concurrency
asyncio
aiohttp>=3.8.0
uvloop>=0.17.0

# Time Series & Trend Analysis
prophet>=1.1.4
statsmodels>=0.14.0
arch>=5.5.0

# Image & Video Processing
pillow>=10.0.0
imageio>=2.31.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
colorama>=0.4.6 